[CHORE] : moved right click menu code from Room.tsx to Member.tsx
