+++ frontend/components/main/member_list/Member.tsx
@@ -3,9 +3,10 @@
 import Image from "next/image";
 import StarIcon from "@mui/icons-material/Star";
 import "@/components/main/member_list/MemberList.css";
-import { useState } from "react";
+import { useState, MouseEvent } from "react";
 import MemberModal from "./MemberModal";
-import { IMember } from "@/components/public/Layout";
+import { IMember } from "@/context/RoomContext";
+import { Menu, MenuItem } from "@mui/material";
 
 export default function Member({
   idx,
@@ -15,14 +16,31 @@ export default function Member({
   person: IMember;
 }) {
   const [openModal, setOpenModal] = useState<boolean>(false);
+  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
 
   const handleOpenModal = () => {
     setOpenModal(true);
   };
-  
+
+  const handleOpenMenu = (
+    e: MouseEvent<HTMLButtonElement, globalThis.MouseEvent>
+  ) => {
+    e.preventDefault();
+    setAnchorEl(e.currentTarget);
+  };
+
+  const handleCloseMenu = () => {
+    setAnchorEl(null);
+  };
+
   return (
     <>
-      <div key={idx} className="membtn" onClick={handleOpenModal}>
+      <div
+        key={idx}
+        className="membtn"
+        onClick={handleOpenModal}
+        onContextMenu={(e) => handleOpenMenu(e)}
+      >
         <div className="memimg">
           <Image src="/seal.png" alt="profile" width={53} height={53} />
         </div>
@@ -31,6 +49,16 @@ export default function Member({
           <StarIcon sx={{ height: "15px", color: "yellow" }} />
         </div>
       </div>
+      <Menu
+        anchorEl={anchorEl}
+        open={Boolean(anchorEl)}
+        onClose={handleCloseMenu}
+      >
+        <MenuItem>Set Admin</MenuItem>
+        <MenuItem>Mute</MenuItem>
+        <MenuItem>Kick</MenuItem>
+        <MenuItem>Ban</MenuItem>
+      </Menu>
       {/* <MemberModal
         openModal={openModal}
         setOpenModal={setOpenModal}
